requirements.txt
openai
python-dotenv
pydantic-ai
absl-py==2.1.0
attrs==24.2.0
cattrs==1.5.0
cloudpickle==3.1.0
filelock==3.16.1
fsspec==2024.10.0
grpcio==1.68.1
gym==0.26.2
gym-notices==0.0.8
h5py==3.12.1
Markdown==3.7
MarkupSafe==3.0.2
mlagents==0.30.0
mlagents_envs==0.30.0
numpy==1.23.5
packaging==24.2
PettingZoo==1.15.0
pillow==11.0.0
protobuf==3.20.3
pypiwin32==223
pywin32==308
PyYAML==6.0.2
six==1.16.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==1.13.1+cu117
typing_extensions==4.12.2
Werkzeug==3.1.3
